backend/requirements.txt
beautifulsoup4==4.13.4
dacite==1.8.1
dask[complete]==2025.1.0
django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
h5py==3.12.1
igraph==0.11.8
loguru==0.7.3
matplotlib==3.10.0
numpy==2.2.2
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyzmq==26.4.0
scipy==1.15.1
seaborn==0.13.2
tqdm==4.67.1
yappi==1.6.10
